Updating IntCodeComputer to have more memory and support longs

+++ AdventOfCode2019/Day2.cs
@@ -6,16 +6,16 @@ namespace AdventOfCode2019
 {
     public class Day2
     {
-        private static readonly int[] Inputs = File.ReadAllText(@"Inputs/Day2.txt").Split(',').Select(int.Parse).ToArray();
+        private static readonly long[] Inputs = File.ReadAllText(@"Inputs/Day2.txt").Split(',').Select(long.Parse).ToArray();
         public void Run()
         {
             Console.WriteLine(PartOne(12, 2));
             PartTwo();
         }
 
-        private int PartOne(int noun, int verb)
+        private long PartOne(int noun, int verb)
         {
-            var inputs = new int[Inputs.Length];
+            var inputs = new long[Inputs.Length];
 
             Array.Copy(Inputs, 0, inputs, 0, Inputs.Length);
